+++ packages/chart/src/index.ts
@@ -2,5 +2,5 @@ export * from './common/config';
 export * from './hooks/useChartDimensions';
 export * from './hooks/useCartesianEncodings';
 export * from './line/LineChart';
-export * from './utils/getInnerGraphDimension';
 export * from './frames/SvgWithAxisFrame';
+export * from './utils/getInnerGraphDimensionAndMargin';