+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a custom effect `useChartDimensions` to calculate the outer and inner dimension of the chart. (#16)
 - Add `useCartesianEncodings` to calculate the processed data and the visual encodings that we need in order to draw the graph. (#16)
 - Add a <SvgWithAxisFrame> component to deal with the size of the chart container, SVG, and the axes that generally used across different charts. (#16)
+- Add `type-check` and `type-check:watch` script in package.json. (#15)
+- Add legend setting in color encoding and show `<Legend>` in `<LineChart>`. (#14)
 - Add `getColorScale` to receive color encoding in charts. (#11)
 - Extract the animation frame controls as a custom effect hook. (#10)
 - Add `tslint-react-hooks` rules to lint React Hooks. (#8)
@@ -29,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Changed
 - Remove `selectors` from `AxisScale` and `ColorScale`. (#16)
 - Modify the config of `tslint` so that it won't continuing warning about the lack of dangling commas in functions. (#16)
+- Fix `HoverLayer` default props. (#14)
+- Fix color too light / deep with ordinal data field. (#14)
 - Rewrite `<HoverLayer>` using Hooks. (#10)
 - Remove `AxisConfig` type. Add new `Encoding` (`AxisEncoding`) and `Scale` (`AxisScale`) type for `getAxisScale`. (#9)
 - Replace `fieldX` `fieldY` `scaleX` `scaleY` with `x` and `y` props in `LineChart`. (#9)